Slight edit
Changed the output of snpExtractConverter.py so the histogram bars
would appear better.
E.g. instead of trying to draw from position 1 to 1 for a mutation at
position 1, it will draw from position 1 to 2 for mutation at position
1. This won't effect position 2 (if there is a mutation) because it
will be 2 to 3 and so on. This makes the bar more visible.